ゲーム画面描画処理をDOMContentLoaded後に実行
