+++ public/game_screen_react.html
@@ -13,6 +13,10 @@
   <script src="https://unpkg.com/react-dom@18/umd/react-dom.production.min.js" crossorigin></script>
   <!-- 外部CSS（必要なら） -->
   <link rel="stylesheet" href="game_screen.css" />
+  <!-- React用コンポーネントを読み込み -->
+  <script defer src="components/Sparkline.js"></script>
+  <script defer src="components/IndicatorCard.js"></script>
+  <script defer src="components/GameScreen.js"></script>
   <!-- React版スクリプト読み込み -->
   <script defer src="game_screen_react.js"></script>
 </head>
